Add more flexibility to the update-aur.sh script
Summary:
This diffs makes the script to update a single package at a time, and
take an optional commit message to replace the default version/release
bump message.

Test Plan:
Comment the final push, then:
Should fail:
  ./update-aur.sh
  ./update-aur.sh this has too many arguments
Should succeed:
  ./update-aur.sh bitcoin-abc
  ./update-aur.sh bitcoin-abc "my commit message"
Check that the commit message is the one passed on the command line.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D7267