tests: add deterministic chain generation unittest fixture
Summary:
This is a backport of [[https://github.com/bitcoin/bitcoin/pull/19806 | core#19806]] [3/8]
https://github.com/bitcoin/bitcoin/pull/19806/commits/31d225274ff1a4b245aea0a69f0e5224b0e64ca2

Note that the hash of the block in `TestChain100Setup` is different because Bitcoin ABC blocks have differences with Bitcoin Core blocks.
I added a test that does nothing but initialize a `TestChain100DeterministicSetup` to demonstrate that the hash in the assertion is correct (changing something in that string makes the test fail).

Test Plan: `ninja check`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D11225