[CI] Move some more path definitions to the Configuration object
Summary:
D7310 made the project root part of the configuration object, as it is
independent of the build context (Teamcity or user machine). The build
directory is also context independent, and the same applies to most of
its subdirectories. Only the artifact directory location needs to be
adapted depending on the context.

This diff moves the build directory path definition to the Configuration
object, as well as its subdirectories. The rational is the same than for
the project root, it allow to derive the paths to construct the build
steps from the configuration.

There is no change in behavior.

Depends on D7310.

Test Plan: Run any CI build.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D7311