[CI] Prevent artifacts from the same diretory to fail the build
Summary:
The `shutil.copytree` method will raise an exception if the destination
already exists, which can easily happen when 2 artifacts are copied to
the same directory. There is a flag to prevent this but unfortunately it
requires python 3.8, so just discard the exception for now.

Test Plan:
Create a collision in the artifact destination directories (currently
there is none by chance) and run the build with the
`build-configurations.py` script. Check the artifacts are copied as
expected and the build succeeds.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D7308