[CI] Clear the artifacts directory before the build
Summary:
The artifact directory is currently cleared after the build, just before
the artifacts are copied.
This will be a problem if we want to place files to the artifact
directory at build time, for example by using an install target.
This diffs clears the artifact directory right before starting the
build.

Test Plan: Run any build, and check the artifacts are copied as previously.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D7309