[CI] Move the project root directory to the Configuration object
Summary:
This is actually part of the Build object, but this is invariant between
the Build types. This also makes it possible for the Configuration
object to derive the path from the project root, which will be required
to generate the build steps from the configuration file.

Test Plan: Run any CI build.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D7310