Fix detection of binary open() calls in the Python encoding linter
Summary:
Currently, the python file encoding linter incorrectly identifies some cases of `open()` needing an encoding explicitly set when they in fact do not.
Reading/writing binary files should not set an encoding, so this false positive lint message will lead the developer to produce incorrect code
if the message is to be believed.

This patch fixes the pattern matching for binary-mode `open()` calls, which will allow the buildbot code to be ported to this repository
without breaking the linter.

Test Plan:
```
arc lint --everything
```

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7326