[land-bot] Extract out conduit token sanitization code
Summary:
The `land-patch.sh` script suffers from a poor design that makes it
not only difficult to test, but also difficult to compose into parts for reuse.
This is made most evident when attempting to modify `land-patch.sh` to support
automated workflows that produce git commits only rather than revisions.

The first step to improving the design is extracting out this critical security
piece so that it can be leverged in multiple places. It also adds test coverage.

This patch does not modify behavior.

Test Plan:
```
cmake -GNinja -DBUILD_SOURCE_CONTROL_TOOLS=ON ..
ninja check-source-control-tools
```

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7352