Add 'sequence' zmq publisher to track all block (dis)connects, mempool deltas
Summary:
Using the zmq notifications to avoid excessive mempool polling can be difficult
given the current notifications available. It announces all transactions
being added to mempool or included in blocks, but announces no evictions
and gives no indication if the transaction is in the mempool or a block.

Block notifications for zmq are also substandard, in that it only announces
block tips, while all block transactions are still announced.

This commit adds a unified stream which can be used to closely track mempool:

1) `getrawmempool` to fill out mempool knowledge
2) if `txhash` is announced, add or remove from set based on add/remove flag
3) if `blockhash` is announced, get block txn list, remove from those transactions local view of mempool
4) if we drop a sequence number, go to (1)

The mempool sequence number starts at the value 1, and
increments each time a transaction enters the mempool,
or is evicted from the mempool for any reason, including
block inclusion. The mempool sequence number is published
via ZMQ for any transaction-related notification.

These features allow for ZMQ/RPC consumer to track mempool
state in a more exacting way, without unnecesarily polling
`getrawmempool`. See interface_zmq.py::test_mempool_sync for
example usage.

This is a backport of core#19572 [2/4]
https://github.com/bitcoin/bitcoin/pull/19572/commits/e76fc2b84d065c9d06010d0a10b316f1f9d36fb9

See for D2304 for the reason of the change in the check-doc.py linter.

Depends on D10303

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D10304