[CI] Migrate the OSX cross build to use no script
Summary:
This diff make the osx CI build use the `cross_build` configuration.
For the build to succeed it is necessary to set the python library path
to use the library built from the depends. This used to be done in the
build-osx script, and has moved to the `build_cmake.sh` script; it is
now always set but will only be used for the osx build.

Depends on D7380.

Test Plan: Run the OSX CI build.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7382