[CI] Don't call the build_cmake.sh script
Summary:
The `build_cmake.sh` script is a wrapper that does 3 things:
 - it cleans the build directory
 - it runs cmake to configure the project
 - it runs ninja on the selected target, all by default

Since the CI build are migrating to descriptive builds, the last 2
pieces are now part of the configuration. The added value of the script
is only the call to `git clean` to clean the build directory.
This diffs make the builds defined fromt the configuration be
independent of the `build_cmake.sh' script. Once all the builds are
migrated, the wrapper can be completely removed.

Depends on D7380.

Test Plan:
  ./contrib/teamcity/build-configurations.py build-asan

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7391