[CMAKE] Make the doc-rpc target actually generate the rpc docs
Summary:
The current target runs the go script, which is not enough to generate
the docs. It also requires a `bitcoind` instance running on regtest, and
the user has to go through the logic of starting the server, running the
target and shutting down the server.
This diff scripts this process so that the `doc-rpc` target handles all
of this and actually generates the RPC docs.

Test Plan:
  ninja doc-rpc

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, deadalnix, jasonbcox

Subscribers: jasonbcox, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7409