[DOC] Fix various doc issues, improve the linter accordingly
Summary:
This diff fixes a few issues:
 1. An obsolete content in the developer notes
 2. An invalid anchor in the Qt README
 3. Links to obsolete directories in the contrib README

The linter is updated to catch the cases 2 and 3. The link regex is
updated to allow for passing the anchors as well as links to directories
(or files without an extension). The gitian builder directory is
excluded as the README contains links relative to project root (starting
with `/`) which will raise false positives in our source tree.

Test Plan:
Read the docs.
  arc lint --everything
Revert the changes to `src/qt/README.md` and `contrib/README.md` and run
`arc lint`. Check the linter finds the issues.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D7436