[CI] Add the failed functional tests tmp directories to artifacts
Summary:
The failed functional tests directories used to be part of the common
artifacts from the configuration file. This template has been removed,
and this diff adds them back as a permanent directory to copy for each
build.

Test Plan:
Edit some functional test to make it fail.
  ./contrib/teamcity/build-configurations.py build-diff
Check the failed tests tmp dirs are copied to the artifacts as expected.

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Subscribers: majcosta

Differential Revision: https://reviews.bitcoinabc.org/D7447