[CMAKE] Fix NSIS installing failing to find the man pages
Summary:
This diff adds back the manpages to the packages (they are no longer part of the default install target) but removes them from the windows packages. The man pages
generation will fail for `bitcoin-qt.exe` on the windows build because
it does not output the help to stdout but creates a window instead.
Since the man pages are not really usable on windows anyway, they are
just removed from the installation package.

Test Plan:
  cmake -GNinja .. \
    -DCMAKE_TOOLCHAIN_FILE=../cmake/platforms/Win64.cmake \
    -DBUILD_BITCOIN_SEEDER=OFF
  ninja package

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7450