Update the AUR packages builds to use the generated man pages
Summary:
The man pages are no longer part of the source tree and need to be
generated for embedding into the packages. Note that `xvfb-run` is used
to make sure `bitcoin-qt --help` can run in a headless environment.

Test Plan: Ran the `update-aur.sh` script in a container for both packages.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7454