Revert "[land-bot] Add buildbot support for requests to land diff patches"
Summary:
This reverts commit 81bfbeea00d1acc839e7f28ac016b9ce43933683.

This reverts D7425 since the rationale behind that design is not strong enough
to warrant the extra complexity.  While a land queue is necessary for human-authored
changes that may be stacked or conflict with each other, this is unlikely to be
true for well thought out automated changes.  For the most part, we can safely
assume automation can run in parallel without causing merge conflicts. And even
if this is not true for special cases, we can work around those for far lower cost
than maintaining a central queue for automated changes.

Test Plan: `ninja check-buildbot`

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7443