[land-bot] Use --nobranch for revisions by default
Summary:
This will allow us to make the assumption that the patch
being landed is on master, whether the patch is part of a revision or not.
This assumption helps keep the patch landing logic as simple as possible
as we begin to use `land-patch.sh` for non-revision patches.

Test Plan:
```
./autopatch.sh --revision Dxxxx
git log    # verify the patch is the latest commit on master
```

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7466