test: Replace use of (dis)?connect_nodes globals
Summary:
This is a 3 commit PR that replaces all uses of `util.(dis)?connect_nodes` with `TestFramework.(dis)?connect_nodes`. This reduces the number of imports by a lot and leave only one single way of doing this operation.

A later scripted-diff commit replaces the majority of uses, which all
follow this pattern:

    `(dis)?connect_nodes(self.nodes[a], self.nodes[b]))`

This commit replaces the few "special cases".

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/19667 | core#19667]] [1/3]
https://github.com/bitcoin/bitcoin/pull/19967/commits/be386840d4a394a1b6221fb7d0fa2b0bc4b1d413

The line touched in rpc_psbt is in a segwit related test that does not exist in Bitcoin ABC.

Test Plan: `ninja check-functional`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D10558