[buildbot] Strip out commonly used separators when detecting backports in diff summaries
Summary:
We sometimes see variants like the following:
* Backport of [[https://github.com/bitcoin/bitcoin/pull/12345 | PR12345]]
* Backport of PR 12345
* Backport of PR#12345
* Backport of PR #12345

Only the first one is automatically linked. We can detect these and both fix and link them
rather than ignoring them. The rationale for fixing them instead of leaving as-is is because
PRxxxxx is easier to grep if they are consistent in all patch summaries.

Other variants of the above are welcome. I simply picked these because we see them the most.

Test Plan: `pytest`

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7489