Decouple archiving release notes from automated commits pipeline
Summary:
This is the last recipe to decouple from the pipeline.
This patch also cleans up the now unnecessary logic regarding COMMIT_TYPE.

The next steps are to further decouple the staged changes from creating the commit
itself, which will make it easier to apply some of the recipes as a form of auto-gen
on top of other patches (ex: bumping package versions ON a version bump, rather than after).

Depends on D7504

Test Plan: `./automated-commits.sh ./patch-recipes/archive-release-notes.sh`

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7505