[CI] Run the circular dependencies linter on CI for each diff
Summary:
Follow-up for the discussion in D7458.

The circular dependencies linter takes too long to run to be integrated
as part of the regular linter test suite (from 2.5s to 6s on my
machines). This diffs adds it to the CI configurations that will run for
every diff. The script is adapted to always run from the project root
and bail out early on failure.

Depends on D7534.

Test Plan:
  ./contrib/teamcity/build-configurations.py lint-circular-dependencies

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D7535