Remove the config.site generation from the depends
Summary:
This template was used to inject some information from the depends build
to the autotools build, and is no longer used.

This diff removes this file and all internal reliance on it, as it was
used to trigger the build. The `config.sub` is now used instead, with
the side effect that it is now properly tracked and will trigger the
depends build when changed.

Test Plan:
Should trigger the build:
  make build-linux64

Should not trigger the build:
  make build-linux64

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7671