Fix the debian package build for Ubuntu Bionic
Summary:
The PPA fails to build the Ubuntu Bionic package at the `dwz` step
(https://manpages.debian.org/testing/debhelper/dh_dwz.1.en.html). We are
installing the files stripped, so `dwz` cannot extract a debug info from
the binaires and fails the build. Note that the version from Ubuntu
Focal correctly detects that there is no debug info and doesn't fail.

This diffs skips the `dwz` step which is not necessary.

Test Plan:
Run the build on Launchpad for Bionic:
https://launchpadlibrarian.net/500034270/buildlog_ubuntu-bionic-amd64.bitcoinabc_0.22.4.2-bionic_BUILDING.txt.gz

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D7682