[SECP256K1] Install a prebuilt cmake version on OSX
Summary:
The homebrew addon is flaky and will fail to update the packages on a
regular basis. The only updated package needed is cmake, so install it
from the prebuilt binaries and cache it to prevent failure. This allow
for calling brew directly on the cached database, solving the flakyness
and saving a lot of build time.

This requires to install openssl explicitely which is obvioulsy (?) no
longer installed since we're not using the homebrew addon anymore. Note
that autotools is not affected by openssl not being installed because it
auto detects it.

Test Plan: Run the Travis build.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7683