Explicitly filter out partial groups when we don't want them
Summary:
```
Instead of hacking OutputGroup::m_ancestors to discourage the inclusion of partial groups via the eligibility filter, add a parameter to the eligibility filter that indicates whether we want to include the group.
Then for those partial groups, don't return them in GroupOutputs if we indicate they aren't desired.
```

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/20040 | core#20040]]:
https://github.com/bitcoin/bitcoin/pull/20040/commits/f6b305273910db0e46798d361413a7e878cb45f7

Depends on D14021.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: PiRK

Differential Revision: https://reviews.bitcoinabc.org/D14022