[buildbot] Introduce an ignore list for known flaky errors
Summary:
The buildbot currently provides periodic reminders when flakiness
arises. This is a hint for devs to investigate the problem and hopefully fix
the flakiness in a reasonable amount of time. However, we cannot always control
how well third-party software performs. In these cases, we may need to wait for
long release cycles or for bug fixes to become available on our current distrobution.

Instead of bringing noise to the dev channel, we can silence known flakiness.
This feature should be used with caution, as ignoring flakiness in our own software
can lead to issues metastasizing.

This patch introduces the first such case, where we ignore a deadlock that occurs
in Wine during cross builds, but only at a frequency of a few times per week.

Test Plan: `pytest`

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7787