Add a docker file to help node deployment
Summary:
This diff adds a `Dockerfile` to deploy a node from a release archive
and a `README` to explain how to use it. This will be used to build the
release images for the docker hub.

The prebuilt archive is used as an input to remove the need for any of
the build dependencies, and keep the image as small as possible.

Test Plan: Read and try the instructions.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7802