Add a docker file to setup a build container
Summary:
This diff adds a docker file to setup a build container. This will make
it easier for users to build the software on systems for which we don't
provide instructions, or don't meet the requirements for some
dependency. This will be made available via docker hub.

The dependencies installation is extracted from the CI setup script, so
there is a single file to maintain and we will know immediately if
something is missing.

Test Plan: Read the instructions and try them.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7815