Fix passing of coin control in SendMoney()
Summary:
D6367 added `coin_control` as a param but did not clean this up
properly to use it.

This is not covered by tests at the moment, but a quick look at D7685 and this
patch should fix the test failure in that draft.

Test Plan: `ninja check check-functional` for sanity

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Subscribers: majcosta

Differential Revision: https://reviews.bitcoinabc.org/D7792