Fix release notes that got automated out of existence
Summary:
This patch loosens the regex that's used to fetch the version number
from the release notes file. The automation was broken due to a new heading that
was added in D7544, but the breakage only occurred after the 0.22.3 release.
The breakage resulted in archiving the release notes to a badly formed file,
effectively erasing release notes as changes were landed via the land bot.

This patch also restores the 0.22.4 release notes that got removed due to the above.

Test Plan:
  ./contrib/source-control-tools/autogen-recipes/50-archive-release-notes.sh

Pre-patch, obvserve that the release notes version is `''`.
Post-patch, the release notes version is matched correctly (and the archiving appropriately skipped).

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D7793