[buildbot] Fix ref master to be full ref name
Summary:
Since the Teamcity upgrade, builds using the default `master` ref
have been soft-failing (build completes successfully, but Teamcity fails it) with this error:
```
The build was triggered in the branch master which does not correspond to any branch monitored by the build VCS roots (the branch is either closed or excluded / not matched by any branch specification). Because of that default branch revisions were set to this build.
```
The only evidence I can find is that the ref name is now considered incorrect, whereas older versions
of Teamcity treated `master` and `refs/heads/master` as the same.

Test Plan: `pytest`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D7827