rpc: calculate fees in getblock using BlockUndo data
Summary:
```
Using block undo data [...] we can now show fee information for each transaction in a block without the need for additional -txindex and/or a ton of costly lookups. For a start we'll add transaction fee information to getblock verbosity level 2. This comes at a negligible speed penalty (<1%).
```

Backport of [[https://github.com/bitcoin/bitcoin/pull/18772 | core#18772]].

Note that we already had a test for the getblock verbose output, so this
one is added using a more explicit name (D2589).

Depends on D12425.

Test Plan:
  ./test/functional/test_runner.py rpc_blockchain

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D12427