util: Add Assume() identity function
Summary:
```
This is needed for #20138. Please refer to the added documentation for
motivation.
```

Backport of [[https://github.com/bitcoin/bitcoin/pull/20255 | core#20255]].

Depends on D9681.

Comes with a linter fix for an edge case where a macro defining
`((void)(val))` would be linted as `(()(val))`.

Ref T1611.

Test Plan:
With and without debug:
  ninja all check

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Maniphest Tasks: T1611

Differential Revision: https://reviews.bitcoinabc.org/D9680