Bump our minimal Qt version to 5.9.5
Summary:
This is the oldest version on the most common linux distributions (from
pkgs.org) for which the build is not already broken:
 - Our current minimum is from Ubuntu Xenial 16.04, which is broken due
to libevent and Boost
 - Next is Debian Stretch which is also broken due to libevent

The picked version is from Ubuntu Bionic 18.04. Other can still use the
static depends for the build, just like before this patch, which has Qt
5.9.7.

We don't test against the 5.5.1 version, and the risk is high that it
gets broken on a regular basis (e.g. D7892, see
https://github.com/bitcoin/bitcoin/pull/14979).

Test Plan:
  ninja all check

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7893